Remove deprecated traits.
