+++ src/ontology/csr/beta.rs
@@ -8,12 +8,7 @@ use std::{
 use graph_builder::{index::Idx, CsrLayout, DirectedCsrGraph, DirectedNeighbors, GraphBuilder};
 
 use crate::{
-    io::{GraphEdge, Relationship, OntologyData},
-    ontology::{
-        HierarchyQueries, HierarchyTraversals, HierarchyWalks, MetadataAware, OntologyTerms,
-    },
-    term::AltTermIdAware,
-    Identified, TermId,
+    io::{GraphEdge, OntologyData, Relationship}, ontology::{HierarchyQueries, HierarchyTraversals, HierarchyWalks, MetadataAware, OntologyTerms}, term::AltTermIdAware, Identified, TermId
 };
 
 /// An ontology backed by a term array and a CSR adjacency matrix.
