+++ src/ontology/csr/mod.rs
@@ -10,8 +10,8 @@
 //!
 //! use ontolius::io::OntologyLoaderBuilder;
 //! use ontolius::ontology::csr::CsrOntology;
-//! use ontolius::term::simple::SimpleMinimalTerm;
 //! use ontolius::ontology::OntologyTerms;
+//! use ontolius::term::simple::SimpleMinimalTerm;
 //!
 //! // Configure the ontology loader to parse Obographs JSON file.
 //! let loader = OntologyLoaderBuilder::new()
@@ -19,15 +19,15 @@
 //!                .build();
 //!
 //! // Load a small Obographs JSON file into `CsrOntology`.
-//! // Use `usize` as ontology graph indices.
+//! // Use `u32` as ontology graph index type.
 //! let path = "resources/hp.small.json.gz";
 //!
-//! /// Use `flate2` to decompress JSON on the fly
+//! /// Use `flate2` to decompress JSON on the fly ...
 //! let reader = GzDecoder::new(BufReader::new(File::open(path).unwrap()));
-//! let ontology: CsrOntology<usize, SimpleMinimalTerm> = loader.load_from_read(reader)
+//! let ontology: CsrOntology<u32, SimpleMinimalTerm> = loader.load_from_read(reader)
 //!                                         .expect("Obographs JSON should be parsable");
 //!
-//! // or do the same using the `MinimalCsrOntology` alias to save some typing:
+//! // ... or do the same using the `MinimalCsrOntology` alias to save some typing:
 //! use ontolius::ontology::csr::MinimalCsrOntology;
 //!
 //! let reader = GzDecoder::new(BufReader::new(File::open(path).unwrap()));
@@ -37,16 +37,10 @@
 //! // Check the number of primary terms
 //! assert_eq!(ontology.len(), 614);
 //! ```
-//!
-//! Check the [`crate::ontology::Ontology`] documentation for more info
-//! regarding the supported functionality.
+
 mod beta;
-mod hierarchy;
-mod ontology;
 
 pub use beta::CsrOntology;
-pub use hierarchy::CsrOntologyHierarchy;
-pub use ontology::CsrOntology as VintageCsrOntology;
 
 use crate::term::simple::{SimpleMinimalTerm, SimpleTerm};
