--- src/prelude.rs
@@ -1,23 +1,0 @@
-//! The list of recommended imports for using the library.
-pub use crate::term::AltTermIdAware;
-pub use crate::term::MinimalTerm;
-pub use crate::term::Term;
-pub use crate::Identified;
-pub use crate::TermId;
-
-pub use crate::hierarchy::AncestorNodes;
-pub use crate::hierarchy::ChildNodes;
-pub use crate::hierarchy::DescendantNodes;
-pub use crate::hierarchy::HierarchyIdx;
-pub use crate::hierarchy::OntologyHierarchy;
-pub use crate::hierarchy::ParentNodes;
-
-pub use crate::io::OntologyData;
-pub use crate::io::OntologyLoader;
-pub use crate::io::OntologyLoaderBuilder;
-
-pub use crate::ontology::HierarchyAware;
-pub use crate::ontology::MetadataAware;
-pub use crate::ontology::Ontology;
-pub use crate::ontology::TermAware;
-pub use crate::ontology::TermIdx;
